refactor: Graph 위치 수정
[#804]
