+++ frontend/src/components/ui/StatisticsGraph/Graph/Bar.stories.tsx
