+++ frontend/src/components/ui/StatisticsGraph/Graph/Bar.tsx
