+++ frontend/src/components/ui/StatisticsGraph/Graph/DayMenus.tsx
