+++ frontend/src/components/ui/StatisticsGraph/Graph/index.tsx
